Update types & pal. for the needs of desktop app
