+++ src/types/operation.js
@@ -1,0 +1,68 @@
+// @flow
+
+/**
+ *
+ *                                 Operation
+ *                                 ---------
+ *
+ */
+
+export type Operation = {
+  // unique identifier (usually hash)
+  id: string,
+
+  // transaction hash
+  hash: string,
+
+  // transaction address
+  address: string,
+
+  // transaction amount in satoshi
+  amount: number,
+
+  // blockHeight allows us to compute the confirmations number (currentBlockHeight - blockHeight)
+  // if null, the operation is not yet on the blockchain
+  blockHeight: number,
+
+  // the account id. available for convenient reason
+  accountId: string,
+
+  // legacy key to store number of confirmations (used by desktop app)
+  // TODO: get rid
+  confirmations: number,
+
+  // --------------------------------------------- specific operation raw fields
+
+  // transaction date
+  date: Date,
+}
+
+export type OperationRaw = {
+  // unique identifier (usually hash)
+  id: string,
+
+  // transaction hash
+  hash: string,
+
+  // transaction address
+  address: string,
+
+  // transaction amount in satoshi
+  amount: number,
+
+  // blockHeight allows us to compute the confirmations number (currentBlockHeight - blockHeight)
+  // if null, the operation is not yet on the blockchain
+  blockHeight: number,
+
+  // the account id. available for convenient reason
+  accountId: string,
+
+  // legacy key to store number of confirmations (used by desktop app)
+  // TODO: get rid
+  confirmations: number,
+
+  // --------------------------------------------- specific operation raw fields
+
+  // transaction date
+  date: string,
+}
